Fix size_t/uint32_t/uint16_t implicit conversions

+++ xoz/arch.h
@@ -3,7 +3,7 @@
 #include <bit>
 
 constexpr uint16_t u16_byteswap(uint16_t x) noexcept {
-    return (x >> 8) | (x << 8);
+    return (uint16_t)((x >> 8) | (x << 8));
 }
 
 constexpr uint32_t u32_byteswap(uint32_t x) noexcept {
